feat(app): compile pkg/app/appdisc under TinyGo (empty-updater on wasm)
appdisc is the next dependency on the path to porting pkg/app/appserver.
Its service-discovery updaters use servicedisc.HTTPClient (net/http) and the
Factory carries *http.Client and *geo.LocationData fields (geo also pulls
net/http), none of which compile on the TinyGo js/wasm target.

A browser visor does not register in service discovery, so the wasm build
simply hands out the existing no-op emptyUpdater — no need to port
servicedisc:
- discovery_manager.go keeps only the Updater interface + emptyUpdater;
  serviceUpdater + PublicVisorUpdater (and the servicedisc import) move to
  discovery_manager_native.go (//go:build !tinygo).
- factory.go keeps the Factory struct + setDefaults, with Client and Geo
  typed `any` (an *http.Client / *geo.LocationData on native) so the shared
  struct is net/http-free. The servicedisc-backed constructors move to
  factory_native.go (recovering the concrete types via asserts);
  factory_tinygo.go returns emptyUpdater for VisorUpdater/AppUpdater
  (PublicVisorUpdater is native-only — no wasm caller references it).

One native caller (visor/init_transport.go) read Factory.Client as
*http.Client; it now recovers it with a comma-ok assert.

Verified: native go build ./..., go vet, appdisc tests, lint unchanged;
`tinygo build -target wasm` of an appdisc importer passes;
`GOOS=js GOARCH=wasm go list -tags tinygo -deps ./pkg/app/appdisc` is
net/http-free.
